Add convert to loot into the npc actor sheet
