+++ templates/tidy-npc-currency-row.html
@@ -19,6 +19,9 @@
         <a style="margin-left: 4px;" class="currency-regenerate rollable" data-action="regenerateCurrency" title="Regnerate Currency">
           <i class="fas fa-redo"></i>
         </a>
+        <a style="margin-left: 4px;" class="currency-loot rollable" data-action="convertToLoot" title="Convert to Loot">
+          <i class="fas fa-shopping-bag"></i>
+        </a>
       </li>
     </ul>
   </div>
